third_party/WebKit/LayoutTests/fast/frames/deep-nested-about-blank.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
window.onload = function() {
    test(function() {
        var doc = document;
        for (i = 0; i < 10; i++) {
            var frame = doc.createElement('iframe');
            frame.src = 'about:blank';
            doc.body.appendChild(frame);
            assert_not_equals(frame.contentDocument, null);
            assert_not_equals(frame.clientHeight, 0);
            doc = frame.contentDocument;
        }
    }, 'Nested about:blank iframes should not be blocked as self-referential iframes.');
}
</script>
